backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai-whisper
ffmpeg-python
certifi
openai
python-dotenv
